Fritzing electronics schematic
This is the electronics circuit for the Arduino piece of the project.
You will need Fritzing (free at http://fritzing.org/) to view the file.

Fritzing has 3 modes. Breadboard, Schematic and PCB.
Once any view is built the other two views are automatically created,
however some tweaking is required to re-route the tracks to make it look
nicer. In this case, the breadboard view is what I used to build the
circuit, the other two views I left alone and are therefore incomplete.
These views are not really needed unless you are planning on printing
your own PCB.